feat: criando view model e configurando RetrofitClient.kt

+++ app/src/main/java/com/curie/curie/data/model/Tarefa.kt
@@ -1,7 +1,10 @@
 package com.curie.curie.data.model
 
+import com.curie.curie.data.model.enums.Prioridade
+import com.curie.curie.data.model.enums.Status
+
 data class Tarefa(
-    val id: Int? = null,
+    val id: Long? = null,
     val userId: Int,
     val nome: String,
     val prazo: String,
